Fix compiler warnings for -Wall -Werror compatibility
This commit addresses three types of compiler warnings:

1. std::ptr_fun deprecation (GCC 13)
   - Replace std::ptr_fun with lambda in psimpl.h
   - Modern C++ lambda is preferred over deprecated adaptor

2. Variable shadowing (MSVC C4457)
   - Rename loop variable 'v' to 'val' in polygon.h
   - Prevents shadowing of function parameter

3. size_t to int conversions (MSVC C4267)
   - Add explicit static_cast<int> for QVector::resize() calls
   - Affects lidar_sensor.cpp, polygon.cpp, triangulator.cpp
   - Qt 5's QVector uses int for size parameters

+++ veranda/Packages/builtin_plugin_logic/pkg-veranda_builtin_shapes/src/triangulator.cpp
@@ -18,7 +18,7 @@ void simplify(Shape &s, const double &crossThreshold)
 QPolygonF simplify(const QPolygonF &p, const double& crossThreshold)
 {
     QPolygonF out;
-    out.resize(p.size());
+    out.resize(static_cast<int>(p.size()));
 
     PolygonIterator inStart((QPolygonF::iterator)p.begin()), inEnd((QPolygonF::iterator)p.end()), outStart((QPolygonF::iterator)out.begin());
     PolygonIterator outEnd = psimpl::simplify_douglas_peucker<2, PolygonIterator, PolygonIterator>
